pyproject.toml
[build-system]
requires = ["setuptools >= 70.0", "autosemver"]
# build-backend = "setuptools.build_meta"

[project]
requires-python = ">=3.12"
dependencies = [
    "autosemver"
]
name = "campbell-control"
dynamic = ["version"]
description = "A python package for controlling Campbell dataloggers over MQTT"

[project.optional-dependencies]
test = ["pytest", "pytest-cov", "parameterized"]
docs = ["sphinx", "sphinx-copybutton", "sphinx-rtd-theme"]
lint = ["ruff"]
paho = ["paho-mqtt"]
aws = ["awscrt"]
all = ["campbell-control[paho,aws]"]
dev = ["campbell-control[test,docs,lint,all]"]

[tool.setuptools.dynamic]
version = { attr = "campbellcontrol.__version__" }


[tool.setuptools.packages.find]
where = ["src"]
include = ["campbellcontrol*"]

[tool.pytest.ini_options]

addopts = "-m 'not hardware' --cov=campbellcontrol"
markers = [
    "hardware: Tests that require a logger connected",
]

filterwarnings = [
    "ignore::DeprecationWarning:autosemver.*:",
    "ignore::DeprecationWarning:pkg_resources.*:",
]

[tool.coverage.run]
omit = ["*__init__.py"]

[tool.ruff]
src = ["src", "tests"]
include = ["src/**.py", "tests/**.py"]
line-length = 120

[tool.ruff.lint]
select = ["F", "E", "W", "A", "PLC", "PLE", "PLW", "I", "N816", "ANN001", "ANN201", "ANN202", "ANN205", "ANN206"]
exclude = ["tests/**.py"]

[tool.ruff.lint.flake8-type-checking]
strict = true